+++ tests/test_checker/test_noqa.py
@@ -14,13 +14,16 @@ https://wemake-python-styleguide.rtfd.io/en/latest/pages/api/contributing.html
 
 import re
 import subprocess
-import types
 from collections import Counter
 from typing import Final
 
 import pytest
 
-from wemake_python_styleguide.compat.constants import PY313, PY315
+from wemake_python_styleguide.compat.constants import (
+    PY313,
+    PY315,
+    make_immutable,
+)
 
 #: Used to find violations' codes in output.
 ERROR_PATTERN = re.compile(r'(WPS\d{3})')
@@ -43,7 +46,7 @@ IGNORED_VIOLATIONS3_13 = ()
 IGNORED_VIOLATIONS3_15 = ()
 
 #: Number and count of violations that would be raised.
-SHOULD_BE_RAISED = types.MappingProxyType(
+SHOULD_BE_RAISED = make_immutable(
     {
         'WPS000': 0,  # logically unacceptable.
         'WPS100': 0,  # logically unacceptable.
@@ -254,6 +257,7 @@ SHOULD_BE_RAISED = types.MappingProxyType(
         'WPS480': 0,  # only triggers on 3.12+
         'WPS481': 10,
         'WPS482': 0,  # enabled only in python 3.15+
+        'WPS483': 0,  # enabled only in python 3.15+
         'WPS500': 1,
         'WPS501': 1,
         'WPS502': 0,  # disabled since 1.0.0
@@ -313,8 +317,9 @@ SHOULD_BE_RAISED = types.MappingProxyType(
 )
 
 #: Number and count of violations that would be raised.
-SHOULD_BE_RAISED3_13 = types.MappingProxyType({'WPS477': 1})
-SHOULD_BE_RAISED3_15 = types.MappingProxyType({'WPS482': 1})
+
+SHOULD_BE_RAISED3_13 = make_immutable({'WPS477': 1})
+SHOULD_BE_RAISED3_15 = make_immutable({'WPS482': 1, 'WPS483': 1})
 
 
 def _assert_errors_count_in_output(